feat: added commands rather than reading args
